requirements.txt
requests==2.32.3
msal==1.32.0
pandas==2.2.3
python-dotenv==1.1.0
